Override much of the build process
Okay some things that needed to be done.

`autogen.sh` wasn't being run in the regular build tools. I don't know
if this is a change to how autoreconf works or what, but feels like
something that needed to be adjusted. I'm sure there's a better way to
do this, but oh well.

When testing, there's two problems.

`dh_auto_test` uses `make check` to run, but since we're not using pip
to install our dependencies and using apt instead, we are beholden to
how debian defines `cram`, which in this case it uses
`cram3`. Annoying since `cram` does not exist in the debian
repository anymore.

The last problem is a hard one to diagnose. `dh` by default likes to
run things in multiple processes. This includes the tests. Within the
tests there is a race condition if run in parallel where a test will
fail with the wrong error because it's trying to read a file it
creates. In this situation it claims the wrong test fails, but I can't
tell if that's because that test secretly expects the previous test to
be successful or if it's just spitting out an error in the middle of
another test run and gets confused.

Either way, this now successfully builds with the appropriate stuff.